batch: Move ownership acquisition helpers
ownership.py defines the persisted ownership model and metadata
serialization.

The scoped acquisition context and detached-copy helper are lifecycle
utilities. Keeping them in the model module made ownership carry buffer
lifetime concerns next to record definitions.

This commit addresses that by moving the acquisition context into
ownership_acquisition and the detached-copy helper into
ownership_detachment, then routing reset-claims through the detached
helper.

Subsequent test commits will update direct helper imports and record the
boundary.
